Solver places rows with only one possible arrangement first
This allows the partial column matcher to do an early exit
if a full row later in the puzzle would otherwise cause
a lot of partial column matchings to be invalid.

This substantially improves the runtime for Puzzle 38 (8sec -> 0.03 sec)
while slowing down the speed of Puzzle 57 (4.6sec -> 6.3 sec).

Go figure.

Other attempted optimizations:
When cloning matrices, do a "JSON.stringify" only once, which can
be "JSON.parse"'d as many times as needed

Inside the solver, use 1 and 0 for cell values so that lines can
easily be summed.