Populate initial puzzle matrix in solver with known marks
A mark is 'known' if it exists in every possible arrangement.

Populating column marks substantially reduces the amount
of row arrangements that need to be checked, because many
of them won't match the partial column arrangements.

Populating row marks allows the partialMatch column checker to
flag more partial column arrangements as impossible.

Also: save off the sums of all the column hints so they
aren't needlessly computed over and over.

All this brings down the speed on the deadly Puzzle 57 to under
three seconds.