Use protractor 5.1
Tell TravisCI to use the latest-ish firefox, because older
firefoxes aren't compatible with newer selenium-webdrivers.

In the TravisCI environment, instead of 'directConnect'
(which fails mysteriously), spawn a selenium server
during the pretest phase.

Also, don't try to check for console error messages in Firefox because
geckodriver seems not to be able to.